Разделил код на файлы
