+++ constants.py
@@ -1,0 +1,5 @@
+FLAG_NOT_SET = 0
+FLAG_ADDED = 1
+FLAG_UNKNOWN = 2
+
+CELL_WITH_MINE = -1
